Finish export of command ids
